+++ index.html
@@ -12,7 +12,7 @@
     <header>
       <h1 class="project-title">2do<span class="box">box</span></h1>
       <input aria-label="title-field" type="text" name="title-field" placeholder="Title" class="title-field">
-      <textarea aria-label="body-field" type="text" name="body-field" class="body-field"  placeholder="body"></textarea>
+      <textarea aria-label="body-field" type="text" name="body-field" class="body-field"  placeholder="task"></textarea>
       <button type="button" name="save" disabled="true" class="save-button">save</button>
     </header>
     <button type="button" name="show-completed" class="show-completed">show completed</button>
