+++ package.json
@@ -23,6 +23,7 @@
   "homepage": "https://github.com/apsitos/2DoBox-Pivot#readme",
   "devDependencies": {
     "babel": "^6.5.2",
+    "babel-core": "^6.18.2",
     "babel-loader": "^6.2.8",
     "babel-preset-es2015": "^6.18.0",
     "chai": "^3.5.0",
