+++ webpack.config.js
@@ -8,5 +8,15 @@ module.exports = {
   output: {
     path: __dirname,
     filename: "[name].bundle.js"
+  },
+  module: {
+    loaders: [
+      { test: /\.js$/, exclude: '/node_modules/', loader: 'babel-loader', query: { presets: ['es2015'] } },
+      { test: /\.css$/, loader: "style!css" },
+      { test: /\.scss$/, loader: "style!css!sass" }
+    ]
+  },
+  resolve: {
+    extensions: ['', '.js', '.json', '.scss', '.css']
   }
 };
